ev_agent_poc/requirements.txt
pandas
streamlit
matplotlib
pyyaml
langchain
ollama